+++ espflash/src/flasher/stubs.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use base64::{engine::general_purpose, Engine as _};
 use serde::{Deserialize, Serialize};
 
@@ -17,6 +19,13 @@ pub struct FlashStub {
     /// Start of data section address
     data_start: u32,
 }
+
+pub(crate) const CHIP_DETECT_MAGIC_REG_ADDR: u32 = 0x40001000;
+pub(crate) const DEFAULT_TIMEOUT: Duration = Duration::from_secs(3);
+pub(crate) const EXPECTED_STUB_HANDSHAKE: &str = "OHAI";
+
+pub(crate) const FLASH_SECTOR_SIZE: usize = 0x1000;
+pub(crate) const FLASH_WRITE_SIZE: usize = 0x400;
 
 // Include stub objects in binary
 const STUB_32: &str = include_str!("../../resources/stubs/stub_flasher_32.toml");
