+++ espflash/src/lib.rs
@@ -37,11 +37,8 @@ pub mod command;
 pub mod connection;
 pub mod elf;
 pub mod error;
-pub mod flash_data;
-#[cfg(feature = "serialport")]
 pub mod flasher;
 pub mod image_format;
-pub mod progress;
 pub mod targets;
 
 /// Logging utilities
