+++ espflash/src/targets/flash_target/mod.rs
@@ -1,12 +1,20 @@
 pub(crate) use self::ram::MAX_RAM_BLOCK_SIZE;
 pub use self::{esp32::Esp32Target, ram::RamTarget};
-#[cfg(feature = "serialport")]
-use crate::{connection::Connection, elf::RomSegment, error::Error, progress::ProgressCallbacks};
+use crate::{connection::Connection, elf::RomSegment, error::Error};
 
 mod esp32;
 mod ram;
 
-#[cfg(feature = "serialport")]
+/// Progress update callbacks
+pub trait ProgressCallbacks {
+    /// Initialize some progress report
+    fn init(&mut self, addr: u32, total: usize);
+    /// Update some progress report
+    fn update(&mut self, current: usize);
+    /// Finish some progress report
+    fn finish(&mut self);
+}
+
 /// Operations for interacting with a flash target
 pub trait FlashTarget {
     /// Begin the flashing operation
